Pulling Councilor Information for IHS Students
The previous request pulled the student id from the registration page, but for some schools including Independence High School, the Student ID is not available from the registration page. This information, though, is accessible through the schedule information. The path for the correct student id without error has been provided.